+++ include/FastaParser/FastaParser.h
@@ -13,8 +13,8 @@ class FastaParser {
     private:
         std::vector<std::pair<std::string, std::string>> sequences;
         float average_L = 0;
-        int max_L = 0;
-        int min_L = INT8_MAX;
+        size_t max_L = 0;
+        size_t min_L = SIZE_MAX;
         
         void parse(std::string path);
